+++ ur_robot_motion_primitives_driver/README.md
@@ -1,4 +1,4 @@
-motion_primitives_ur_driver_pkg
+ur_robot_motion_primitives_driver
 ==========================================
 
 Driver package to control ur robot using motion primitives like MOVEJ (PTP), MOVEL (LIN) and MOVEC (CIRC)
@@ -22,20 +22,20 @@ ros2 launch ur_robot_driver ur_control.launch.py ur_type:=ur5e robot_ip:=172.20.
 ```
 ## Launch motion_primitives_ur_driver
 ```
-ros2 launch motion_primitives_ur_driver_pkg motion_primitive_controller_ur.launch.py ur_type:=ur5e robot_ip:=172.20.0.2 launch_rviz:=true
+ros2 launch ur_robot_motion_primitives_driver motion_primitive_controller_ur.launch.py ur_type:=ur5e robot_ip:=172.20.0.2 launch_rviz:=true
 ```
 ## Publish dummy commands
 ### Single motion primitive
 ```
-ros2 run motion_primitives_ur_driver_pkg send_single_dummy_motion_primitive.py
+ros2 run ur_robot_motion_primitives_driver send_single_dummy_motion_primitive.py
 ```
 ### Multiple motion primitives after checking status=2 (success)
 ```
-ros2 run motion_primitives_ur_driver_pkg send_multiple_dummy_motion_primitives_after_checking_status.py
+ros2 run ur_robot_motion_primitives_driver send_multiple_dummy_motion_primitives_after_checking_status.py
 ```
 ### Multiple motion primitives without checking status
 ```
-ros2 run motion_primitives_ur_driver_pkg send_multiple_dummy_motion_primitives.py
+ros2 run ur_robot_motion_primitives_driver send_multiple_dummy_motion_primitives.py
 ```
 ## Publish stop motion command 
 ```
