+++ ur_robot_motion_primitives_driver/scripts/send_multiple_dummy_motion_primitives_after_checking_status.py
