+++ ur_robot_motion_primitives_driver/urdf/motion_primitive_ur.ros2_control.xacro
@@ -2,7 +2,7 @@
 <robot xmlns:xacro="http://www.ros.org/wiki/xacro">
 
   <!-- <xacro:include filename="$(find ur_description)/urdf/inc/ur_joint_control.xacro" /> -->
-  <xacro:include filename="$(find motion_primitives_ur_driver_pkg)/urdf/ur_joint_control.xacro" />
+  <xacro:include filename="$(find ur_robot_motion_primitives_driver)/urdf/ur_joint_control.xacro" />
   <xacro:include filename="$(find ur_description)/urdf/inc/ur_sensors.xacro" />
 
   <xacro:macro name="motion_primitive_ur_ros2_control" params="
@@ -46,7 +46,7 @@
           <param name="calculate_dynamics">true</param>
         </xacro:if>
         <xacro:unless value="${use_mock_hardware}">
-          <plugin>motion_primitives_ur_driver_pkg/MotionPrimitivesUrDriver</plugin>
+          <plugin>ur_robot_motion_primitives_driver/MotionPrimitivesUrDriver</plugin>
           <param name="robot_ip">${robot_ip}</param>
           <param name="script_filename">${script_filename}</param>
           <param name="output_recipe_filename">${output_recipe_filename}</param>
