+++ ur_robot_motion_primitives_driver/urdf/motion_primitive_ur.urdf.xacro
@@ -9,7 +9,7 @@
   <!-- ros2 control include -->
   <!-- <xacro:include filename="$(find ur_robot_driver)/urdf/ur.ros2_control.xacro" /> -->
   <xacro:include filename="$(find ur_description)/urdf/inc/ur_transmissions.xacro" />
-  <xacro:include filename="$(find motion_primitives_ur_driver_pkg)/urdf/motion_primitive_ur.ros2_control.xacro"/>
+  <xacro:include filename="$(find ur_robot_motion_primitives_driver)/urdf/motion_primitive_ur.ros2_control.xacro"/>
 
   <!-- possible 'ur_type' values: ur3, ur3e, ur5, ur5e, ur10, ur10e, ur16e, ur20 -->
   <!-- the default value should raise an error in case this was called without defining the type -->
